[REFAC]: rule.Parser — RuleLookup interface 추상화 + extract.go 분리 (이슈 #463)
(1) RuleLookup interface 추출:
  Parser.resolver 가 concrete *Resolver 대신 interface 의존으로 전환.
  - 신규 RuleLookup interface (ResolveByURL 메소드 1개)
  - *Resolver 가 자연스럽게 RuleLookup 만족 — 호출자 (main.go / 기존 테스트) 변경 없음
  - 단위 테스트에서 RuleLookup mock 주입 가능 — Redis / cache 의존 제거

(2) parser.go 슬림화 — 추출 헬퍼 → extract.go 분리:
  parser.go (349줄) → engine 진입점만 유지하도록 227줄로 축소.
  분리 대상:
    - hasRequiredSelector / validateRaw
    - extractField / extractFieldFromSelection / extractFromSelection
    - extractFieldMulti / extractValue / extractDate
    - absoluteURL / defaultDateLayouts
  → internal/processor/parser/rule/extract.go (149줄) 신설.

  parser.go 는 ParsePage / ParseLinks 본체 + 인터페이스 conformance 검증만 보유 —
  responsability 명확화 (SRP).
